Reorder all unittests under two packages of keras and pytorch packages
