+++ .github/workflows/run_keras_tests.yml
@@ -29,7 +29,6 @@ jobs:
         # CPU environment (https://github.com/tensorflow/tensorflow/issues/41718).
         # For this reason, if we run them in such an environment, we need to run them first non-parallel separately.
         run: |
-          python -m unittest discover tests/keras_tests/pruning_tests -v
-          python -m unittest discover tests/keras_tests/non_parallel_tests -v
-          for script in tests/xquant_tests/keras_tests tests/keras_tests/exporter_tests tests/keras_tests/feature_networks_tests tests/keras_tests/graph_tests tests/keras_tests/layer_tests; do python -m unittest discover $script -v & pids+=($!); done; for pid in ${pids[@]}; do wait $pid || exit 1; done          
+          python -m unittest discover tests/keras_tests -v
+
