+++ .github/workflows/run_pytorch_tests.yml
@@ -26,7 +26,6 @@ jobs:
           pip install torch==${{ inputs.torch-version }} torchvision onnx onnxruntime
       - name: Run unittests
         run: |
-          python -m unittest discover tests/xquant_tests/pytorch_tests -v
           python -m unittest discover tests/pytorch_tests -v
