+++ tests/external_tests/keras_tests/models_tests/test_networks_runner_float.py
