+++ tests/keras_tests/data_generation_tests/test_scheduler_step.py
