+++ tests/pytorch_tests/xquant_tests/test_similarity_functions.py
